+++ index.js
@@ -71,6 +71,27 @@ module.exports = {
         'EVENTS',
       ],
     }],
+    // Vue 组件的属性排序
+    'vue/order-in-components': ['error', {
+      // 排序顺序
+      order: [
+        'name',
+        'key',
+        'el',
+        'functional',
+        ['props', 'propsData'],
+        'setup',
+        ['provide', 'inject'],
+        'data',
+        'computed',
+        ['template', 'render'],
+        'methods',
+        'LIFECYCLE_HOOKS',
+        'watch',
+        'components',
+        'mixins',
+      ],
+    }],
     // 在单行元素的内容前后需要换行符
     'vue/singleline-html-element-content-newline': 'off',
     // 在 if 语句后必须换行